FRONT/API NFT --Added NFT CRUD operations !! Each NFT Saves as a record in the database + link to the token in Polygon --Added authorization through wallets --Added functionality to view all nfts in profile, add to favorites, view favorites, view profile, edit profile, set avatar and banner, and bio

+++ pinterest_front/src/pages/Auth/RegisterForm.jsx
@@ -1,7 +1,5 @@
 ﻿import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { useRegisterMutation, useGoogleAuthMutation } from '../../../store/Auth/AuthApi.js';
-import { setCredentials } from '../../../store/slices/AuthSlice.js';
+import { usePinterestAuth } from '@/hooks/usePinterestAuth';
 import { Button, Typography, useTheme, Box } from '@mui/material';
 import { useNavigate } from "react-router";
 import InputField from '../../components/ui/Auth/InputField';
@@ -13,28 +11,19 @@ const RegisterForm = () => {
     const theme = useTheme();
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    const [dateOfBirth, setDateOfBirth] = useState('');
-    const [register, { isLoading, error }] = useRegisterMutation();
-    const [googleAuth, { isLoading: isGoogleLoading }] = useGoogleAuthMutation();
-    const dispatch = useDispatch();
+    const [confirmPassword, setConfirmPassword] = useState('');
+    const { register, isLoading } = usePinterestAuth();
     const navigate = useNavigate();
     const [showPassword, setShowPassword] = useState(false);
 
     const handleSubmit = async (e) => {
         e.preventDefault();
         try {
-            const response = await register({
-                email,
-                password,
-                dateOfBirth,
-            }).unwrap();
-
-            dispatch(setCredentials({
-                user: { email: email },
-                accessToken: response.payload.accessToken
-            }));
-
-            localStorage.setItem('userPassword', password);
+            if (password !== confirmPassword) {
+                console.error('Паролі не збігаються');
+                return;
+            }
+            await register(email, password);
             localStorage.setItem('isNewUser', 'true');
             navigate('/');
         } catch (err) {
